+++ BackEnd/park_management/src/main/java/park/management/com/vn/model/request/BranchStaffRequest.java
@@ -14,8 +14,9 @@ public class BranchStaffRequest {
     private String role;
 
     @NotNull
-    private Integer userId;
+    private Long userId;
 
     @NotNull
-    private Integer parkBranchId;
+    private Long parkBranchId;
+
 }
