+++ BackEnd/park_management/src/main/java/park/management/com/vn/model/response/RoleResponse.java
@@ -3,6 +3,7 @@ package park.management.com.vn.model.response;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
+import park.management.com.vn.entity.Role;
 
 import java.time.LocalDateTime;
 
@@ -10,7 +11,7 @@ import java.time.LocalDateTime;
 @Setter
 @Getter
 public class RoleResponse {
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
     private LocalDateTime createdAt;
