+++ BackEnd/park_management/src/main/java/park/management/com/vn/repository/UserRepository.java
@@ -1,14 +1,17 @@
 package park.management.com.vn.repository;
 
 import java.util.Optional;
+
+import jakarta.validation.constraints.NotNull;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import park.management.com.vn.entity.Users;
 
 @Repository
-public interface UserRepository extends JpaRepository<Users, Integer> {
+public interface UserRepository extends JpaRepository<Users, Long> {
 
   Optional<Users> findByUsername(String username);
 
   boolean existsByEmail(String email);
+
 }
